refactor(v0.81.0): session_checkpoint + transcript → core/runtime_state/ 신설 (의존성 정리 A4, 5/5)
새 사이클 A4. v0.52에서 deferred됐던 33개 ignore_imports 중 추가 6개 정리
(단일 PR 최대 감소).

신설 패키지: core/runtime_state/
- __init__.py     11L  cross-layer session state primitives
- session_checkpoint.py  239L  SessionState + SessionCheckpoint (atomic JSON, /resume 백킹)
- transcript.py          314L  SessionTranscript 대화 로거 + cleanup_old_transcripts

이전: core/cli/{session_checkpoint,transcript}.py
v0.40 historical placement. 사용자 facing CLI 의존성 0. 3 layer (cli, agent,
server)에서 모두 호출되는 cross-layer state primitives.

호출자 업데이트 (14곳, 14 파일):
- core/server/ipc_server/poller.py
- core/agent/loop/{loop,_lifecycle}.py
- core/cli/commands/session.py
- core/cli/cmd_lifecycle.py
- tests/{conftest,test_session_checkpoint,test_phase3_ipc,test_session_manager,test_session_resume,test_session_transcript,test_transcript}.py

pyproject.toml ignore_imports 6 entries 제거 (Agent + Server 양 contracts):
- core.agent.loop.loop -> core.cli.{session_checkpoint,transcript}
- core.agent.loop._lifecycle -> core.cli.session_checkpoint
- core.server.ipc_server.poller -> core.cli.session_checkpoint

28 → 22 ignore_imports 남음 (단일 PR 최대 감소).

Quality Gates:
- ruff/ruff-format/mypy 0 errors (330 source files, +1)
- pytest 4344 passed, 1 skipped, 24 deselected — 변동 없음
- session/transcript 테스트 91개 통과
- import-linter 4/4 contracts 깨짐 0
- E2E Cowboy Bebop A (68.4) 변동 없음

5 PR 진척:
- A1 redaction → utils       ✅ v0.78.0
- A2 bash_tool → agent       ✅ v0.79.0
- A3 project_detect → utils  ✅ v0.80.0
- A4 (this) ckpt + transcript → runtime_state/ ✅ v0.81.0
- A5 startup → lifecycle (마지막 1건)
