+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "geode"
-version = "0.80.0"
+version = "0.81.0"
 description = "GEODE — 범용 자율 실행 에이전트"
 license = "Apache-2.0"
 readme = "README.md"
@@ -99,9 +99,6 @@ forbidden_modules = ["core.cli", "core.server"]
 ignore_imports = [
     # Tier 3 #7 — loop.py was split into core/agent/loop/ package
     "core.agent.loop.loop -> core.cli.commands",
-    "core.agent.loop.loop -> core.cli.session_checkpoint",
-    "core.agent.loop.loop -> core.cli.transcript",
-    "core.agent.loop._lifecycle -> core.cli.session_checkpoint",
     "core.agent.approval -> core.cli",
     "core.agent.worker -> core.cli.tool_handlers",
 ]
@@ -123,15 +120,11 @@ ignore_imports = [
     "core.lifecycle.bootstrap -> core.cli.startup",
     "core.agent.worker -> core.cli.tool_handlers",
     # Tier 3 #7 — loop.py was split into core/agent/loop/ package
-    "core.agent.loop.loop -> core.cli.session_checkpoint",
-    "core.agent.loop.loop -> core.cli.transcript",
     "core.agent.loop.loop -> core.cli.commands",
-    "core.agent.loop._lifecycle -> core.cli.session_checkpoint",
     "core.agent.approval -> core.cli",
     "core.server.ipc_server.poller -> core.cli",
     "core.server.ipc_server.poller -> core.cli.session_state",
     "core.server.ipc_server.poller -> core.cli.startup",
-    "core.server.ipc_server.poller -> core.cli.session_checkpoint",
 ]
 
 [[tool.importlinter.contracts]]
